INTERLOK-541 - Added example generation. INTERLOK-541 #time 1h
Added tests that create examples
Removed @XStreamImplicit from the properties KeyValuePairSet (because
it will look messy with just <key-value-pair> tags inside the XML.
Added Password.decode() use for Passwords in SSL which means a minor
change to the interface to throw an exception with corresponding change
to AdvancedSolaceImpl.
Verified that it publishes manually.